chore(mypy): zera dívida de tipos na camada de dados e torna mypy bloqueante
A checagem de tipos nunca rodava de fato no CI; ao habilitá-la surgiram 203
achados (o output vinha truncado por um crash de encoding, mascarando o total).
A maioria (≈160) é atrito de stub do Cartopy GeoAxes no glue de renderização
(attr-defined em .add_feature/.set_extent/.gridlines) — não são bugs.

Estratégia: type-checagem estrita e BLOQUEANTE na camada de dados/lógica;
isenção declarativa do glue de render Cartopy.

Camada de dados (corrigida de verdade):
- ecmwf.py 87→0: anota VARIABLE_REGISTRY e os dicts de `.sel` como dict[str, Any]
  (mata index/arg-type/assignment), guarda de None no meteograma, np.asarray em
  retornos, best_diff_sec como float, ignore direcionado no overload de datetime64.
- MODOS e _*_style_state anotados (drawing_panel); no-any-return resolvido em
  symbols/base, point_symbols, zcit_axis, loczcit_pa_engine, blocking_engine;
  guardas de None em sst e data_service; okta como float em stations; _MEIPASS
  (PyInstaller) e render() vs QWidget.render com type:ignore direcionado.
- pyproject: disable_error_code=["import-untyped"] (requests/markdown sem stubs).

Isenção (pyproject [tool.mypy.overrides], ignore_errors): map_canvas,
charts/synoptic, charts/interactive, main_window — render Cartopy.

mypy agora bloqueia no CI. ruff limpo, 385 testes passam, 5 pulados.

+++ cartomet_br/symbols/point_symbols.py
@@ -38,7 +38,7 @@ class _CompoundArtist:
 def _symbol_size(ax, frac: float = 0.035) -> float:
     """Calcula tamanho do símbolo baseado na extensão horizontal do mapa."""
     xlim = ax.get_xlim()
-    return abs(xlim[1] - xlim[0]) * frac
+    return float(abs(xlim[1] - xlim[0]) * frac)
 
 
 def _make_circle_path(cx: float, cy: float, radius: float, n_pts: int = 24) -> Path:
